move Prismic models around
